Add UnitStep ....
Also, refactor mathics.eval.arithmetic to remove eval functions
belonging to mathics.builtin.arithfns, and mathics.builtin.numericfns
