+++ mathics/eval/arithfns/__init__.py
@@ -1,0 +1,3 @@
+"""
+Module tracking eval functions under mathics.builtin.arithfns
+"""
